Proper data structure between filtering and data provider is now in use

+++ GPUPatternMining/GPUPatternMining.Contract/Enity/DataFeed.h
@@ -12,7 +12,7 @@ struct Coords
 
 struct DataFeed
 {
-	short* type;
+	int* type;
 	int* instanceId;
 	Coords* xy;
 };
